Fix HTML around excerpt
Problems:
1) if post begins with some non-standard tag, e.g. <blogquote> and
excerpt is cut before it is closed, that excerpt as well as meta
data for the post is quoted.
2) if <a>tag is opened but not closed within the excerpt, that tag
will continue to be present further down the content, resulting in
unwanted behaviour.

Solution: striping HTML tags.

The <p> is added as it was there (as posts in most of the cases
began with that tag), but with stripping the HTML tags they are
removed as well. On the side note - that <p> tag wasn't closed as
well.

Wrapping the excerpt with that tag brings back the behaviour
as well as fixes it being closed.