Use protocol-related URL
Till the domain will redirect to https (by default), this trick
needs to be used to load all assets from same protocol.

Right now, if you enter the blog using secured connection (https)
some resources will be blocked, as the base URL is hardcoded
and uses http.

This is considered anti-pattern, and when the redirection to https
will be done permanently, this value (for url in config) should be
hardcoded to https.

See: https://www.paulirish.com/2010/the-protocol-relative-url/